# openapi.yaml
openapi: 3.0.3
info:
  title: Tzav Rishon API
  description: API for IDF Tzav Rishon psychometric test preparation platform
  version: 1.0.0
  contact:
    name: Tzav Rishon Platform
servers:
  - url: http://localhost:8080/api/v1
    description: Development server

paths:
  /auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/logout:
    post:
      summary: Logout user
      tags: [Auth]
      responses:
        '200':
          description: Logged out successfully

  /practice/start:
    post:
      summary: Start practice session
      tags: [Practice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [VERBAL_ANALOGY, SHAPE_ANALOGY, INSTRUCTIONS_DIRECTIONS, QUANTITATIVE]
      responses:
        '200':
          description: Session started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeSessionResponse'

  /practice/{sessionId}/questions:
    get:
      summary: Get practice questions
      tags: [Practice]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'

  /practice/{sessionId}/answer:
    post:
      summary: Submit practice answer
      tags: [Practice]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Answer result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'

  /exam/start:
    post:
      summary: Start exam attempt
      tags: [Exam]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Exam started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamAttemptResponse'

  /exam/{attemptId}/section/current:
    get:
      summary: Get current exam section
      tags: [Exam]
      security:
        - cookieAuth: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current section details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentSectionResponse'

  /progress/summary:
    get:
      summary: Get progress summary
      tags: [Progress]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Progress summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressSummaryResponse'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tzav_auth

  schemas:
    AuthResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        guestId:
          type: string
          format: uuid

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        isPremium:
          type: boolean

    Question:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        format:
          type: string
        promptText:
          type: string
        promptImageUrl:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'

    QuestionOption:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        imageUrl:
          type: string
        optionOrder:
          type: integer

    PracticeSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        type:
          type: string
        questionsAvailable:
          type: integer
        limitReached:
          type: boolean

    SubmitAnswerRequest:
      type: object
      properties:
        questionId:
          type: string
          format: uuid
        textAnswer:
          type: string
        selectedOptionId:
          type: string
          format: uuid
        timeMs:
          type: integer

    AnswerResponse:
      type: object
      properties:
        correct:
          type: boolean
        explanation:
          type: string

    ExamAttemptResponse:
      type: object
      properties:
        attemptId:
          type: string
          format: uuid
        sections:
          type: array
          items:
            type: object

    CurrentSectionResponse:
      type: object
      properties:
        sectionId:
          type: string
          format: uuid
        type:
          type: string
        remainingTimeSeconds:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'

    ProgressSummaryResponse:
      type: object
      properties:
        totalAttempts:
          type: integer
        overallAccuracy:
          type: number
        avgTimePerQuestionMs:
          type: integer
        improvementPercent:
          type: number