# server/src/main/java/com/tzavrishon/domain/QuestionOption.java
package com.tzavrishon.domain;

import jakarta.persistence.*;
import java.util.UUID;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Table(name = "question_options")
@Data
@NoArgsConstructor
public class QuestionOption {
  @Id
  @GeneratedValue(strategy = GenerationType.UUID)
  private UUID id;

  @ManyToOne(fetch = FetchType.LAZY)
  @JoinColumn(name = "question_id", nullable = false)
  private Question question;

  @Column(columnDefinition = "TEXT")
  private String text;

  @Column(name = "image_url", columnDefinition = "TEXT")
  private String imageUrl;

  @Column(name = "is_correct")
  private Boolean isCorrect = false;

  @Column(name = "option_order", nullable = false)
  private Integer optionOrder;
}